+++ src/app/(core)/teachers/programmes/[slug]/[tab]/Components/ProgrammeView.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { OakMaxWidth, OakTabs } from "@oaknational/oak-components";
-import { useEffect, useMemo, useState } from "react";
+import { useEffect, useState } from "react";
 import { notFound, usePathname, useSearchParams } from "next/navigation";
 
 import {
@@ -32,15 +32,10 @@ import {
   CurriculumDownloadsTierSubjectProps,
   CurriculumUnitsFormattedData,
 } from "@/pages-helpers/curriculum/docx/tab-helpers";
-import { useFilters } from "@/hooks/useFilters";
 import {
   CurriculumSelectionSlugs,
   CurriculumSelectionTitles,
 } from "@/utils/curriculum/slugs";
-import {
-  CurriculumFilters,
-  OnChangeCurriculumFilters,
-} from "@/utils/curriculum/types";
 import { ProgrammePageHeaderCMS } from "@/common-lib/cms-types/programmePage";
 import {
   CurriculumOverviewSanityData,
@@ -49,7 +44,7 @@ import {
 import type { Ks4Option } from "@/node-lib/curriculum-api-2023/queries/curriculumPhaseOptions/curriculumPhaseOptions.schema";
 import { resolveOakHref } from "@/common-lib/urls";
 import { validateSearchParams } from "@/utils/validateProgrammePageSearchParams";
-import { getDefaultFilter } from "@/utils/curriculum/filtering";
+import { useBrowseFilters } from "@/context/BrowseFilters";
 import { useTeacherBrowseAnalytics } from "@/context/TeacherBrowseAnalytics/TeacherBrowseAnalyticsProvider";
 import { ComponentType } from "@/browser-lib/avo/Avo";
 
@@ -66,7 +61,6 @@ export type ProgrammePageProps = {
   tabSlug: TabSlug;
   ks4Options: Ks4Option[];
   ks4OptionFilterDimensions: Record<string, Ks4OptionFilterDimension>;
-  initialFilter?: CurriculumFilters;
   featureFlags: Record<string, boolean>;
   implementationGuides: ImplementationGuides | null;
   fileSizes: ProgrammeDownloadsProps["fileSizes"];
@@ -85,7 +79,6 @@ export const ProgrammeView = ({
   subjectPhaseSlug,
   ks4Options,
   ks4OptionFilterDimensions,
-  initialFilter,
   featureFlags,
   implementationGuides,
   fileSizes,
@@ -100,33 +93,10 @@ export const ProgrammeView = ({
   const { subjectTitle, phaseTitle, examboardTitle } =
     curriculumSelectionTitles;
 
-  const defaultFilter = useMemo(() => {
-    return getDefaultFilter(curriculumUnitsFormattedData);
-  }, [curriculumUnitsFormattedData]);
-
-  const [filters, setFilters] = useFilters(defaultFilter, initialFilter);
-
-  const { programmeRefined, curriculumResourcesAccessed } =
-    useTeacherBrowseAnalytics((store) => store.track);
-
-  const onChangeFilters: OnChangeCurriculumFilters = ({
-    newFilters,
-    filterType,
-    filterValue,
-  }) => {
-    setFilters(newFilters);
-
-    if (!filterType) {
-      return;
-    }
-
-    programmeRefined({
-      componentType: "filter_link",
-      activeFilters: newFilters,
-      filterType,
-      filterValue,
-    });
-  };
+  const { filters } = useBrowseFilters();
+  const { curriculumResourcesAccessed } = useTeacherBrowseAnalytics(
+    (store) => store.track,
+  );
 
   const schoolYear = filters.years.find(
     (year) => validatedParams?.years === year,
@@ -236,8 +206,6 @@ export const ProgrammeView = ({
         curriculumCMSInfo={curriculumCMSInfo}
         curriculumDownloadsTabData={curriculumDownloadsTabData}
         mvRefreshTime={mvRefreshTime}
-        filters={filters}
-        setFilters={onChangeFilters}
         ks4Options={ks4Options}
         ks4OptionFilterDimensions={ks4OptionFilterDimensions}
         implementationGuides={implementationGuides}
@@ -255,8 +223,6 @@ const TabContent = ({
   curriculumCMSInfo,
   curriculumDownloadsTabData,
   mvRefreshTime,
-  filters,
-  setFilters,
   ks4Options,
   ks4OptionFilterDimensions,
   implementationGuides,
@@ -274,8 +240,6 @@ const TabContent = ({
       <UnitSequenceView
         curriculumSelectionSlugs={curriculumSelectionSlugs}
         curriculumUnitsFormattedData={curriculumUnitsFormattedData}
-        filters={filters}
-        setFilters={setFilters}
         ks4Options={ks4Options}
         ks4OptionFilterDimensions={ks4OptionFilterDimensions}
       />