+++ src/app/(core)/teachers/programmes/[slug]/[tab]/page.tsx
@@ -32,7 +32,11 @@ import withPageErrorHandling, {
 } from "@/hocs/withPageErrorHandling";
 import { resolveOakHref } from "@/common-lib/urls";
 import { getSubjectPhaseSlug } from "@/components/TeacherComponents/helpers/getSubjectPhaseSlug";
-import { resolveFilterFromSearchParams } from "@/utils/curriculum/filtering";
+import {
+  getDefaultFilter,
+  resolveFilterFromSearchParams,
+} from "@/utils/curriculum/filtering";
+import { BrowseFiltersProvider } from "@/context/BrowseFilters";
 import { redirectProgrammeSlugIfNeeded } from "@/utils/integratedJourney/legacyProgrammeUnitsRedirect";
 import { cacheData, CURRICULUM_API_CACHE_TAG } from "@/node-lib/cache";
 import CMSClient from "@/node-lib/cms";
@@ -239,6 +243,7 @@ const InnerProgrammePage = async (props: AppPageProps<ProgrammePageParams>) => {
     formatCurriculumUnitsData(curriculumUnitsData);
 
   // Resolve filter server-side from URL search params
+  const defaultFilter = getDefaultFilter(curriculumUnitsFormattedData);
   const resolvedFilter = resolveFilterFromSearchParams(
     curriculumUnitsFormattedData,
     searchParams,
@@ -322,7 +327,6 @@ const InnerProgrammePage = async (props: AppPageProps<ProgrammePageParams>) => {
     ks4OptionFilterDimensions,
     curriculumDownloadsTabData,
     mvRefreshTime,
-    initialFilter: resolvedFilter,
     fileSizes,
     featureFlags: {
       "implementation-guides": isImplementationGuidesEnabled,
@@ -342,8 +346,13 @@ const InnerProgrammePage = async (props: AppPageProps<ProgrammePageParams>) => {
       programmeState={programmeState}
       accessLevel="programme"
     >
-      <ProgrammeView {...results} />
-      {isEnabled && <LayoutPreviewControls />}
+      <BrowseFiltersProvider
+        defaultFilter={defaultFilter}
+        initialFilter={resolvedFilter}
+      >
+        <ProgrammeView {...results} />
+        {isEnabled && <LayoutPreviewControls />}
+      </BrowseFiltersProvider>
     </TeacherBrowseAnalyticsStoreProvider>
   );
 };